Added more classes and fixed sensor support
Some scripts were changed to use more classes in the code to make it tidier and easier to read. Sensor support was mostly rewritten and the circuit controlling the sensors was reconstructed.  A log file and a writing system was added to produce training data for the AI. On a separate note, a chassis of the car was found and motors were mounted at the car. We are planning to use octocouplers to control the analog input for the motors. The car weights around 580± grams at this point.

Co-Authored-By: nuh <80310745+notnaki@users.noreply.github.com>