Motor control added(partial, needs some work)
Through a bridge, added motor control. It can be controlled using the pins 17,22,25 for motor 1 and 37,35,33 for motor 2. But one problem with the setup is the total flow not being enough to supply the raspberry pi. We will try to solve this problem by using batteries on the bridge as power so that we can power the motors. This will add around 100g to the total weight with the batteries and we will have a slight issue actually supplying the batteries. Also, the wiring for the motor2 may be incorrect.